move base, serial and arduino into utilities folder
